+++ src/components/modal/ModalEditLocationPrice.tsx
@@ -1,29 +1,13 @@
-import React, { useEffect, useState } from 'react'
-import {
-  Button,
-  Form,
-  Input,
-  InputNumber,
-  Modal,
-  Radio,
-  RadioChangeEvent,
-  Select,
-  Space,
-  Table,
-} from 'antd'
+/* eslint-disable react/jsx-key */
+import React, { useState } from 'react'
+import { Button, Form, Input, InputNumber, Modal, Radio, Table } from 'antd'
 import FooterPage from '../footer/FooterPage'
-import {
-  AllLocatePriceEntity,
-  LocationPricePageEntity,
-  PricePlantsEntity,
-  UpdateLocationPrice,
-  UpdateLocationPriceList,
-  UpdateLocationPriceList_INIT,
-  UpdateLocationPrice_INIT,
-} from '../../entities/LocationPrice'
+import { AllLocatePriceEntity, UpdateLocationPriceList } from '../../entities/LocationPrice'
 import { LocationPriceDatasource } from '../../datasource/LocationPriceDatasource'
 import { SearchOutlined } from '@ant-design/icons'
 import { color } from '../../resource'
+import SwitchButton from '../button/SwitchButton'
+import { useLocalStorage } from '../../hook/useLocalStorage'
 
 const _ = require('lodash')
 const { Map } = require('immutable')
@@ -48,6 +32,10 @@ const ModalEditLocationPrice: React.FC<ModalLocationPriceProps> = ({
   const [getSearch, setGetSearch] = useState<AllLocatePriceEntity[]>()
   const [saveBtnDisable, setBtnSaveDisable] = useState<boolean>(false)
   const [searchText, setSearchText] = useState<string>('')
+  const [selectPrice, setSelectPrice] = useState<any>('equal')
+  const [type, setType] = useState<any>('SPRAY')
+  const [equalPrice, setEqualPrice] = useState<number>()
+  const [profile] = useLocalStorage('profile', [])
 
   const searchPlants = async () => {
     await LocationPriceDatasource.getPrice(data.province_name, searchText).then((res) => {
@@ -62,7 +50,65 @@ const ModalEditLocationPrice: React.FC<ModalLocationPriceProps> = ({
     setSearchText(search.target.value)
   }
   const handelCallBack = async () => {
-    callBack(locationPrice)
+    if (selectPrice === 'equal') {
+      const equalData: any = {}
+      equalData.provinceId = data?.province_id
+      if (type === 'SOW') {
+        equalData.priceSow = equalPrice
+      } else if (type === 'SPRAY') {
+        equalData.price = equalPrice
+      }
+      equalData.updateBy = `${profile.firstname} ${profile.lastname}`
+      callBack(equalData)
+    } else {
+      callBack(locationPrice)
+    }
+  }
+  const onChangeType = (e: any) => {
+    setType(e.target.value)
+  }
+
+  const handleEqualPrice = (e: any) => {
+    setEqualPrice(e.target.value)
+  }
+  const renderPriceColumn = (value: any, row: any, index: number) => {
+    const handleItemClick = (indexData: number, newItem: number | null) => {
+      const newItems = [...locationPrice]
+      const dataChange: UpdateLocationPriceList[] = []
+      const price_id = Map(dataChange).set('location_price_id', row.id)
+      let price: any
+      if (type === 'SPRAY') {
+        price = Map(price_id.toJS()).set('price', newItem || 0)
+      } else {
+        price = Map(price_id.toJS()).set('priceSow', newItem || 0)
+      }
+
+      newItems[indexData] = price.toJS()
+      setLocationPrice(newItems)
+      setBtnSaveDisable(!newItem)
+    }
+
+    const inputNumberProps = {
+      key: type === 'SOW' ? row.price_sow : row.price,
+      style: {
+        width: 200,
+        color: (type === 'SOW' ? row.price_sow : row.price) === null ? 'grey' : 'black',
+      },
+      defaultValue:
+        (type === 'SOW' ? row.price_sow : row.price) !== null
+          ? type === 'SOW'
+            ? row.price_sow
+            : row.price
+          : 'ยังไม่มีข้อมูล',
+      step: '0.01',
+      onChange: (event: number | null) => handleItemClick(index, event),
+      stringMode: true,
+      addonAfter: 'บาท',
+    }
+
+    return {
+      children: <InputNumber {...inputNumberProps} />,
+    }
   }
   const columns = [
     {
@@ -78,38 +124,10 @@ const ModalEditLocationPrice: React.FC<ModalLocationPriceProps> = ({
     },
     {
       title: 'ราคา',
-      dataIndex: 'price',
-      key: 'price',
+      dataIndex: 'prices',
+      key: 'prices',
       width: '50%',
-      render: (value: any, row: any, index: number) => {
-        const handleItemClick = (indexData: number, newItem: number) => {
-          const newItems = [...locationPrice]
-          const dataChange: UpdateLocationPriceList[] = []
-          const price_id = Map(dataChange).set('location_price_id', row.id)
-          const price = Map(price_id.toJS()).set('price', newItem)
-          newItems[indexData] = price.toJS()
-          setLocationPrice(newItems)
-
-          {
-            !newItem ? setBtnSaveDisable(true) : setBtnSaveDisable(false)
-          }
-        }
-        return {
-          children: (
-            <>
-              <InputNumber
-                key={row.price}
-                style={{ width: 200 }}
-                defaultValue={row.price}
-                step='0.01'
-                onChange={(event) => handleItemClick(index, event)}
-                stringMode
-                addonAfter='บาท'
-              />
-            </>
-          ),
-        }
-      },
+      render: renderPriceColumn,
     },
   ]
 
@@ -143,45 +161,72 @@ const ModalEditLocationPrice: React.FC<ModalLocationPriceProps> = ({
               โปรดตรวจสอบราคาการฉีดพ่นก่อนที่จะกดบันทึกเสมอ เพราะอาจจะทำให้ราคาฉีดพ่น ของอำเภอ
               และตำบลทั้งหมดเปลี่ยนแปลงตาม อาจจะส่งผลต่อการจ้างงานในระบบ
             </span>
-            <div className='col-lg-10 pt-3'>
-              <Input
-                allowClear
-                prefix={<SearchOutlined style={{ color: color.Disable }} />}
-                placeholder='ค้นหาชื่อพืช'
-                onChange={changeTextSearch}
-              />
+            <div className='pt-3'>
+              <SwitchButton label1={type} label2={type} onClick={onChangeType} />
             </div>
-            <div className='col-lg-2 pt-3'>
-              <Button
-                style={{
-                  borderColor: color.Success,
-                  borderRadius: '5px',
-                  color: color.secondary2,
-                  backgroundColor: color.Success,
-                  padding: 6,
-                  paddingTop: 4,
-                }}
-                onClick={searchPlants}
-              >
-                ค้นหาข้อมูล
-              </Button>
-            </div>
+            <label className='pt-2'>
+              ราคา<span style={{ color: color.Error }}>*</span>
+            </label>
+            <Radio.Group
+              name='radiogroup'
+              defaultValue={'equal'}
+              onChange={(e) => setSelectPrice(e.target.value)}
+            >
+              <Radio value='equal'>ใช้ราคาเท่ากัน</Radio>
+              <Radio value='customize'>กำหนดรายพืช</Radio>
+            </Radio.Group>
+            {selectPrice === 'equal' ? (
+              <div className='pt-3'>
+                <Input
+                  placeholder='กรอกราคา'
+                  suffix='บาท'
+                  onChange={(e) => handleEqualPrice(e)}
+                  autoComplete='off'
+                />
+              </div>
+            ) : (
+              <>
+                <div className='col-lg-10 pt-3'>
+                  <Input
+                    allowClear
+                    prefix={<SearchOutlined style={{ color: color.Disable }} />}
+                    placeholder='ค้นหาชื่อพืช'
+                    onChange={changeTextSearch}
+                  />
+                </div>
+                <div className='col-lg-2 pt-3 pb-4'>
+                  <Button
+                    style={{
+                      borderColor: color.Success,
+                      borderRadius: '5px',
+                      color: color.secondary2,
+                      backgroundColor: color.Success,
+                      padding: 6,
+                      paddingTop: 4,
+                    }}
+                    onClick={searchPlants}
+                  >
+                    ค้นหาข้อมูล
+                  </Button>
+                </div>
+                {getSearch ? (
+                  <Table
+                    dataSource={getSearch?.map((x) => x.plants)[0]}
+                    columns={columns}
+                    pagination={false}
+                    scroll={{ x: 0, y: 300 }}
+                  />
+                ) : (
+                  <Table
+                    dataSource={data.plants}
+                    columns={columns}
+                    pagination={false}
+                    scroll={{ x: 0, y: 300 }}
+                  />
+                )}
+              </>
+            )}
           </div>
-          {getSearch ? (
-            <Table
-              dataSource={getSearch?.map((x) => x.plants)[0]}
-              columns={columns}
-              pagination={false}
-              scroll={{ x: 0, y: 300 }}
-            />
-          ) : (
-            <Table
-              dataSource={data.plants}
-              columns={columns}
-              pagination={false}
-              scroll={{ x: 0, y: 300 }}
-            />
-          )}
         </Form>
       </Modal>
     </>