+++ src/pages/droner/IndexDroner.tsx
@@ -420,7 +420,7 @@ function IndexDroner() {
       title: () => {
         return (
           <div style={{ display: 'flex', gap: 8, alignItems: 'center' }}>
-            อัพเดตล่าสุด
+            อัปเดตล่าสุด
             <div
               style={{
                 display: 'flex',