+++ src/pages/highlightNews/HighlightNewsPage.tsx
@@ -337,7 +337,7 @@ function HighlightNewsPage() {
       title: () => {
         return (
           <div style={{ display: 'flex', gap: 8, alignItems: 'center' }}>
-            กดอ่าน
+            กดลิ้งค์
             <div
               style={{
                 display: 'flex',