+++ src/pages/redeem/DetailFarmerRedeem.tsx
@@ -96,7 +96,7 @@ const DetailFarmerRedeem = () => {
 
   const columeHis = [
     {
-      title: 'วันที่อัพเดท',
+      title: 'วันที่อัปเดต',
       dataIndex: 'dateTime',
       key: 'dateTime',
       render: (value: any, row: any, index: number) => {
@@ -135,7 +135,7 @@ const DetailFarmerRedeem = () => {
       },
     },
     {
-      title: 'ผู้ใช้ที่อัพเดต',
+      title: 'ผู้ใช้ที่อัปเดต',
       dataIndex: 'updateBy',
       key: 'updateBy',
       render: (value: any, row: any, index: number) => {