+++ src/pages/setting/targetSpray/IndexTargetSpray.tsx
@@ -115,7 +115,7 @@ function IndexTargetSpray() {
       title: () => {
         return (
           <div style={{ display: 'flex', gap: 8, alignItems: 'center' }}>
-            อัพเดตล่าสุด
+            อัปเดตล่าสุด
             <div
               style={{
                 display: 'flex',