+++ src/pages/task/AdminTask/IndexAdminTask.tsx
@@ -1178,7 +1178,7 @@ const IndexAdminTask = () => {
   )
   const columns = [
     {
-      title: 'วันที่อัพเดต',
+      title: 'วันที่อัปเดต',
       dataIndex: 'createdAt',
       key: 'createdAt',
       render: (value: any, row: any, index: number) => {
@@ -1248,7 +1248,7 @@ const IndexAdminTask = () => {
       },
     },
     {
-      title: 'ผู้ใช้ที่อัพเดต',
+      title: 'ผู้ใช้ที่อัปเดต',
       dataIndex: 'createdBy',
       key: 'createdBy',
       render: (value: any, row: any, index: number) => {